+++ src/github.com/matrix-org/dendrite/clientapi/auth/storage/accounts/accounts_table.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package storage
+package accounts
 
 import (
 	"database/sql"
