+++ src/github.com/matrix-org/dendrite/clientapi/auth/types/device.go
@@ -14,16 +14,10 @@
 
 package types
 
-import (
-	"github.com/matrix-org/gomatrixserverlib"
-)
-
-// Account represents a Matrix account on this home server.
-type Account struct {
-	UserID     string
-	Localpart  string
-	ServerName gomatrixserverlib.ServerName
-	// TODO: Other flags like IsAdmin, IsGuest
-	// TODO: Devices
-	// TODO: Associations (e.g. with application services)
+// Device represents a client's device (mobile, web, etc)
+type Device struct {
+	ID          string
+	UserID      string
+	AccessToken string
+	// TODO: display name, last used timestamp, keys, etc
 }
